# openapi.yaml
openapi: 3.0.3
info:
  title: VGMdb API
  description: |
    API for scraping and parsing video game music database information from VGMdb.net

    **Enhanced TypeScript/Node.js Implementation**

    This API provides comprehensive access to VGMdb data including:
    - Album information with ratings, related albums, stores, and cover galleries
    - Artist profiles with discography, websites, and social media
    - Product/game information with franchise hierarchies and release tables
    - Event, organization, and release data
    - Search and recent updates

    All endpoints support both JSON and YAML response formats.
  version: 2.0.0
  contact:
    name: VGMdb API
    url: https://vgmdb.net
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:9990
    description: Local development server
  - url: https://vgmdb.info
    description: Production server (if deployed)

tags:
  - name: Info
    description: Album, artist, product, event, org, and release information
  - name: Lists
    description: Browse lists of albums, artists, products, orgs, and events
  - name: Search
    description: Search across all content types
  - name: Recent
    description: Recent updates and additions

paths:
  /hello:
    get:
      summary: Health check
      description: Test if the server is running
      tags:
        - Info
      responses:
        '200':
          description: Server is running
          content:
            text/plain:
              schema:
                type: string
                example: Hello from VGMdb TypeScript Node.js!

  /album/{id}:
    get:
      summary: Get album information
      description: |
        Retrieve detailed information about a specific album including:
        - Basic info (name, catalog, release date, media format)
        - Cover art (thumb, small, full)
        - Track listings with composers, arrangers, performers
        - **Enhanced**: Ratings, votes, related albums, stores, websites, cover gallery
      tags:
        - Info
      parameters:
        - name: id
          in: path
          required: true
          description: Album ID
          schema:
            type: string
          example: '79'
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Album information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumInfo'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  /artist/{id}:
    get:
      summary: Get artist information
      description: |
        Retrieve detailed information about an artist including:
        - Basic info (name, birthdate, birthplace)
        - **Enhanced**: Sex, type, alias relationships, death date
        - **Enhanced**: Discography, featured works, websites, Twitter
      tags:
        - Info
      parameters:
        - name: id
          in: path
          required: true
          description: Artist ID
          schema:
            type: string
          example: '291'
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Artist information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistInfo'
        '404':
          $ref: '#/components/responses/NotFound'

  /product/{id}:
    get:
      summary: Get product/game information
      description: |
        Retrieve detailed information about a game or product including:
        - Basic info (name, type, release date)
        - Franchises and organizations
        - **Enhanced**: Franchise hierarchy (superproduct, subproducts, titles)
        - **Enhanced**: Release table with platforms and regions
        - Related albums
      tags:
        - Info
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
          example: '189'
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Product information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductInfo'
        '404':
          $ref: '#/components/responses/NotFound'

  /event/{id}:
    get:
      summary: Get event information
      description: Retrieve information about an event (convention, concert, etc.)
      tags:
        - Info
      parameters:
        - name: id
          in: path
          required: true
          description: Event ID
          schema:
            type: string
          example: '138'
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Event information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventInfo'
        '404':
          $ref: '#/components/responses/NotFound'

  /org/{id}:
    get:
      summary: Get organization information
      description: Retrieve information about a record label or publisher
      tags:
        - Info
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID
          schema:
            type: string
          example: '67'
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Organization information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgInfo'
        '404':
          $ref: '#/components/responses/NotFound'

  /release/{id}:
    get:
      summary: Get release information
      description: Retrieve information about a specific release version
      tags:
        - Info
      parameters:
        - name: id
          in: path
          required: true
          description: Release ID
          schema:
            type: string
          example: '1234'
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Release information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseInfo'
        '404':
          $ref: '#/components/responses/NotFound'

  /albumlist/{id}:
    get:
      summary: Get album list
      description: Get a list of albums by letter or number
      tags:
        - Lists
      parameters:
        - name: id
          in: path
          required: true
          description: List ID (A1, A2, ..., Z, 0-9)
          schema:
            type: string
          example: 'F'
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Album list retrieved successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /artistlist/{id}:
    get:
      summary: Get artist list
      description: Get a list of artists by letter or number
      tags:
        - Lists
      parameters:
        - name: id
          in: path
          required: true
          description: List ID (A1, A2, ..., Z, 0-9)
          schema:
            type: string
          example: 'N'
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Artist list retrieved successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /productlist/{id}:
    get:
      summary: Get product list
      description: Get a list of products by letter or number
      tags:
        - Lists
      parameters:
        - name: id
          in: path
          required: true
          description: List ID (A1, A2, ..., Z, 0-9)
          schema:
            type: string
          example: 'F'
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Product list retrieved successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /orglist:
    get:
      summary: Get organization list
      description: Get a list of all organizations
      tags:
        - Lists
      parameters:
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Organization list retrieved successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /eventlist:
    get:
      summary: Get event list
      description: Get a list of all events
      tags:
        - Lists
      parameters:
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Event list retrieved successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /search:
    get:
      summary: Search
      description: Search for albums, artists, products, and organizations
      tags:
        - Search
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
          example: 'final fantasy'
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Search results retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
        '400':
          description: Missing required parameter 'q'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /recent/{type}:
    get:
      summary: Get recent updates
      description: Get recent updates for a specific type
      tags:
        - Recent
      parameters:
        - name: type
          in: path
          required: false
          description: Type of updates (defaults to 'albums')
          schema:
            type: string
            enum:
              - albums
              - media
              - tracklists
              - scans
              - artists
              - products
              - labels
              - links
              - ratings
            default: albums
          example: 'albums'
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Recent updates retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecentUpdates'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    FormatParam:
      name: format
      in: query
      description: Response format
      schema:
        type: string
        enum:
          - json
          - yaml
        default: json

  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Resource not found'

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Internal server error'

    ServiceUnavailable:
      description: VGMdb.net is temporarily unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'vgmdb.net is temporarily unavailable'

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message

    Names:
      type: object
      description: Multi-language names
      additionalProperties:
        type: string
      example:
        en: 'Final Fantasy VIII Original Soundtrack'
        ja: 'ファイナルファンタジーVIII オリジナル・サウンドトラック'

    MetaInfo:
      type: object
      properties:
        added_date:
          type: string
        added_user:
          type: string
        edited_date:
          type: string
        edited_user:
          type: string
        visitors:
          type: integer
        ttl:
          type: integer

    AlbumInfo:
      type: object
      properties:
        name:
          type: string
        names:
          $ref: '#/components/schemas/Names'
        catalog:
          type: string
        release_date:
          type: string
        media_format:
          type: string
        classification:
          type: string
        picture_thumb:
          type: string
        picture_small:
          type: string
        picture_full:
          type: string
        publisher:
          type: object
        organizations:
          type: array
          items:
            type: object
        composers:
          type: array
          items:
            type: string
        arrangers:
          type: array
          items:
            type: string
        performers:
          type: array
          items:
            type: string
        lyricists:
          type: array
          items:
            type: string
        discs:
          type: array
          items:
            type: object
        notes:
          type: string
          description: '✨ Enhanced: Album notes'
        rating:
          type: number
          description: '✨ Enhanced: User rating'
        votes:
          type: integer
          description: '✨ Enhanced: Number of votes'
        categories:
          type: array
          items:
            type: string
          description: '✨ Enhanced: Full category list'
        products:
          type: array
          items:
            type: object
          description: '✨ Enhanced: Related products'
        platforms:
          type: array
          items:
            type: string
          description: '✨ Enhanced: Gaming platforms'
        related:
          type: array
          items:
            type: object
          description: '✨ Enhanced: Related albums'
        stores:
          type: array
          items:
            type: object
          description: '✨ Enhanced: Online stores'
        websites:
          type: object
          description: '✨ Enhanced: Categorized websites'
        covers:
          type: array
          items:
            type: object
          description: '✨ Enhanced: Cover gallery'
        meta:
          $ref: '#/components/schemas/MetaInfo'

    ArtistInfo:
      type: object
      properties:
        name:
          type: string
        names:
          $ref: '#/components/schemas/Names'
        sex:
          type: string
          enum: [male, female]
          description: '✨ Enhanced: Artist gender'
        type:
          type: string
          description: '✨ Enhanced: Artist type (Individual/Unit/Alias)'
        alias_of:
          type: object
          description: '✨ Enhanced: Alias relationship'
        name_real:
          type: string
          description: '✨ Enhanced: Real name'
        name_trans:
          type: string
          description: '✨ Enhanced: Name transliteration'
        deathdate:
          type: string
          description: '✨ Enhanced: Death date'
        picture_full:
          type: string
        picture_small:
          type: string
        birth_place:
          type: string
        birthdate:
          type: string
        aliases:
          type: array
        notes:
          type: string
        discography:
          type: array
          items:
            type: object
          description: '✨ Enhanced: Complete discography'
        featured_on:
          type: array
          items:
            type: object
          description: '✨ Enhanced: Featured works'
        websites:
          type: object
          description: '✨ Enhanced: Categorized websites'
        twitter_names:
          type: array
          items:
            type: string
          description: '✨ Enhanced: Twitter handles'
        meta:
          $ref: '#/components/schemas/MetaInfo'

    ProductInfo:
      type: object
      properties:
        name:
          type: string
        names:
          $ref: '#/components/schemas/Names'
        type:
          type: string
        picture_full:
          type: string
        picture_small:
          type: string
        description:
          type: string
        release_date:
          type: string
        franchises:
          type: array
          items:
            type: object
        organizations:
          type: array
          items:
            type: object
        superproduct:
          type: object
          description: '✨ Enhanced: Parent franchise'
        subproducts:
          type: array
          items:
            type: object
          description: '✨ Enhanced: Child products'
        titles:
          type: array
          items:
            type: object
          description: '✨ Enhanced: Franchise titles'
        releases:
          type: array
          items:
            type: object
          description: '✨ Enhanced: Release table with platforms/regions'
        albums:
          type: array
          items:
            type: object
        websites:
          type: object
        meta:
          $ref: '#/components/schemas/MetaInfo'

    EventInfo:
      type: object
      properties:
        name:
          type: string
        startdate:
          type: string
        enddate:
          type: string
        notes:
          type: string
        releases:
          type: array
          items:
            type: object
        meta:
          $ref: '#/components/schemas/MetaInfo'

    OrgInfo:
      type: object
      properties:
        name:
          type: string
        names:
          $ref: '#/components/schemas/Names'
        type:
          type: string
        region:
          type: string
        picture_full:
          type: string
        picture_small:
          type: string
        staff:
          type: array
          items:
            type: object
        description:
          type: string
        releases:
          type: array
          items:
            type: object
        websites:
          type: object
        meta:
          $ref: '#/components/schemas/MetaInfo'

    ReleaseInfo:
      type: object
      properties:
        name:
          type: string
        name_real:
          type: string
        type:
          type: string
        picture_full:
          type: string
        picture_small:
          type: string
        products:
          type: array
          items:
            type: object
        catalog:
          type: string
        upc:
          type: string
        release_type:
          type: string
        platform:
          type: string
        region:
          type: string
        release_date:
          type: string
        release_albums:
          type: array
          items:
            type: object
        product_albums:
          type: array
          items:
            type: object
        meta:
          $ref: '#/components/schemas/MetaInfo'

    SearchResults:
      type: object
      properties:
        query:
          type: string
        sections:
          type: object
          additionalProperties:
            type: object

    RecentUpdates:
      type: object
      properties:
        sections:
          type: array
          items:
            type: object
        updates:
          type: array
          items:
            type: object