# src/index.ts
import express, { NextFunction, Request, Response } from 'express';
import * as yaml from 'js-yaml';
import * as path from 'path';
import swaggerUi from 'swagger-ui-express';
import YAML from 'yamljs';
import { fetchAlbumPage, parseAlbumPage } from './parsers/album';
import { fetchAlbumlistPage, parseAlbumlistPage } from './parsers/albumlist';
import { fetchArtistPage, parseArtistPage } from './parsers/artist';
import { fetchArtistlistPage, parseArtistlistPage } from './parsers/artistlist';
import { fetchEventPage, parseEventPage } from './parsers/event';
import { fetchEventlistPage, parseEventlistPage } from './parsers/eventlist';
import { fetchOrgPage, parseOrgPage } from './parsers/org';
import { fetchOrglistPage, parseOrglistPage } from './parsers/orglist';
import { fetchProductPage, parseProductPage } from './parsers/product';
import {
  fetchProductlistPage,
  parseProductlistPage,
} from './parsers/productlist';
import { parseRecentPage } from './parsers/recent';
import { fetchReleasePage, parseReleasePage } from './parsers/release';
import { parseSearchPage } from './parsers/search';
import { fetchPage } from './utils/fetch';

const app = express();
const PORT = process.env.PORT || 9990;
const DEFAULT_CORS_HEADERS =
  'Origin, User-Agent, If-Modified-Since, Cache-Control';

// Load OpenAPI specification
const openapiPath = path.join(__dirname, '..', 'openapi.yaml');
const swaggerDocument = YAML.load(openapiPath);

// Helper function to send formatted response
function sendFormattedResponse(
  res: Response,
  data: unknown,
  format: string
): void {
  res.setHeader('Cache-Control', 'max-age=86400,public');

  // Default to JSON unless explicitly requested as YAML
  if (format === 'yaml') {
    res.setHeader('Content-Type', 'application/x-yaml');
    res.send(yaml.dump(data));
  } else {
    res.setHeader('Content-Type', 'application/json');
    res.json(data);
  }
}

// Helper function to handle errors
function handleError(error: unknown, res: Response, context: string): void {
  console.error(`Error ${context}:`, error);

  if (error instanceof Error && error.message.includes('503')) {
    res.status(503).json({ error: 'vgmdb.net is temporarily unavailable' });
  } else {
    res.status(500).json({ error: 'Internal server error' });
  }
}

// CORS middleware
app.use((req: Request, res: Response, next: NextFunction) => {
  res.setHeader('Access-Control-Allow-Origin', '*');
  res.setHeader('Access-Control-Allow-Methods', 'GET, HEAD, OPTIONS');
  const headers =
    req.headers['access-control-request-headers'] || DEFAULT_CORS_HEADERS;
  res.setHeader('Access-Control-Allow-Headers', headers);
  next();
});

// Handle OPTIONS requests
app.options('*', (req: Request, res: Response) => {
  res.status(200).send('Of course, this API is free for everyone!');
});

// Swagger UI documentation
app.use('/api-docs', swaggerUi.serve);
app.get(
  '/api-docs',
  swaggerUi.setup(swaggerDocument, {
    customSiteTitle: 'VGMdb API Documentation',
    customCss: '.swagger-ui .topbar { display: none }',
    swaggerOptions: {
      persistAuthorization: true,
      displayRequestDuration: true,
      filter: true,
      tryItOutEnabled: true,
    },
  })
);

// Redirect root to API docs
app.get('/', (_req: Request, res: Response) => {
  res.redirect('/api-docs');
});

// Hello endpoint
app.get('/hello', (req: Request, res: Response) => {
  res.send('Hello from VGMdb TypeScript Node.js!');
});

// Album endpoint
app.get('/album/:id', async (req: Request, res: Response) => {
  try {
    const { id } = req.params;
    const format = (req.query.format as string) || '';

    const html = await fetchAlbumPage(id);
    const albumInfo = parseAlbumPage(html);

    if (!albumInfo) {
      return res.status(404).json({ error: 'Album not found' });
    }

    sendFormattedResponse(res, albumInfo, format);
  } catch (error) {
    handleError(error, res, 'fetching album');
  }
});

// Artist endpoint
app.get('/artist/:id', async (req: Request, res: Response) => {
  try {
    const { id } = req.params;
    const format = (req.query.format as string) || '';

    const html = await fetchArtistPage(id);
    const artistInfo = parseArtistPage(html);

    if (!artistInfo) {
      return res.status(404).json({ error: 'Artist not found' });
    }

    sendFormattedResponse(res, artistInfo, format);
  } catch (error) {
    handleError(error, res, 'fetching artist');
  }
});

// Product endpoint
app.get('/product/:id', async (req: Request, res: Response) => {
  try {
    const { id } = req.params;
    const format = (req.query.format as string) || '';

    const html = await fetchProductPage(id);
    const productInfo = parseProductPage(html);

    if (!productInfo) {
      return res.status(404).json({ error: 'Product not found' });
    }

    sendFormattedResponse(res, productInfo, format);
  } catch (error) {
    handleError(error, res, 'fetching product');
  }
});

// Event endpoint
app.get('/event/:id', async (req: Request, res: Response) => {
  try {
    const { id } = req.params;
    const format = (req.query.format as string) || '';

    const html = await fetchEventPage(id);
    const eventInfo = parseEventPage(html);

    if (!eventInfo) {
      return res.status(404).json({ error: 'Event not found' });
    }

    sendFormattedResponse(res, eventInfo, format);
  } catch (error) {
    handleError(error, res, 'fetching event');
  }
});

// Organization endpoint
app.get('/org/:id', async (req: Request, res: Response) => {
  try {
    const { id } = req.params;
    const format = (req.query.format as string) || '';

    const html = await fetchOrgPage(id);
    const orgInfo = parseOrgPage(html);

    if (!orgInfo) {
      return res.status(404).json({ error: 'Organization not found' });
    }

    sendFormattedResponse(res, orgInfo, format);
  } catch (error) {
    handleError(error, res, 'fetching organization');
  }
});

// Albumlist endpoint
app.get('/albumlist/:id', async (req: Request, res: Response) => {
  try {
    const { id } = req.params;
    const format = (req.query.format as string) || '';

    const html = await fetchAlbumlistPage(id);
    const albumlistInfo = parseAlbumlistPage(html);

    if (!albumlistInfo) {
      return res.status(404).json({ error: 'Album list not found' });
    }

    sendFormattedResponse(res, albumlistInfo, format);
  } catch (error) {
    handleError(error, res, 'fetching album list');
  }
});

// Artistlist endpoint
app.get('/artistlist/:id', async (req: Request, res: Response) => {
  try {
    const { id } = req.params;
    const format = (req.query.format as string) || '';

    const html = await fetchArtistlistPage(id);
    const artistlistInfo = parseArtistlistPage(html);

    if (!artistlistInfo) {
      return res.status(404).json({ error: 'Artist list not found' });
    }

    sendFormattedResponse(res, artistlistInfo, format);
  } catch (error) {
    handleError(error, res, 'fetching artist list');
  }
});

// Productlist endpoint
app.get('/productlist/:id', async (req: Request, res: Response) => {
  try {
    const { id } = req.params;
    const format = (req.query.format as string) || '';

    const html = await fetchProductlistPage(id);
    const productlistInfo = parseProductlistPage(html);

    if (!productlistInfo) {
      return res.status(404).json({ error: 'Product list not found' });
    }

    sendFormattedResponse(res, productlistInfo, format);
  } catch (error) {
    handleError(error, res, 'fetching product list');
  }
});

// Orglist endpoint
app.get('/orglist', async (req: Request, res: Response) => {
  try {
    const format = (req.query.format as string) || '';

    const html = await fetchOrglistPage();
    const orglistInfo = parseOrglistPage(html);

    if (!orglistInfo) {
      return res.status(404).json({ error: 'Organization list not found' });
    }

    sendFormattedResponse(res, orglistInfo, format);
  } catch (error) {
    handleError(error, res, 'fetching organization list');
  }
});

// Eventlist endpoint
app.get('/eventlist', async (req: Request, res: Response) => {
  try {
    const format = (req.query.format as string) || '';

    const html = await fetchEventlistPage();
    const eventlistInfo = parseEventlistPage(html);

    if (!eventlistInfo) {
      return res.status(404).json({ error: 'Event list not found' });
    }

    sendFormattedResponse(res, eventlistInfo, format);
  } catch (error) {
    handleError(error, res, 'fetching event list');
  }
});

// Release endpoint
app.get('/release/:id', async (req: Request, res: Response) => {
  try {
    const { id } = req.params;
    const format = (req.query.format as string) || '';

    const html = await fetchReleasePage(id);
    const releaseInfo = parseReleasePage(html);

    if (!releaseInfo) {
      return res.status(404).json({ error: 'Release not found' });
    }

    sendFormattedResponse(res, releaseInfo, format);
  } catch (error) {
    handleError(error, res, 'fetching release');
  }
});

// Search endpoint
app.get('/search', async (req: Request, res: Response) => {
  try {
    const query = (req.query.q as string) || '';
    const format = (req.query.format as string) || '';

    if (!query) {
      return res.status(400).json({ error: 'Query parameter q is required' });
    }

    const url = `https://vgmdb.net/search?q=${encodeURIComponent(query)}`;
    const html = await fetchPage(url);
    const searchInfo = parseSearchPage(html);

    if (!searchInfo) {
      return res.status(404).json({ error: 'No results found' });
    }

    sendFormattedResponse(res, searchInfo, format);
  } catch (error) {
    handleError(error, res, 'performing search');
  }
});

// Recent endpoint
app.get('/recent/:type?', async (req: Request, res: Response) => {
  try {
    const type = req.params.type || 'albums';
    const format = (req.query.format as string) || '';

    const url = `https://vgmdb.net/db/recent.php?do=view_${type}`;
    const html = await fetchPage(url);
    const recentInfo = parseRecentPage(html);

    if (!recentInfo) {
      return res.status(404).json({ error: 'Recent updates not found' });
    }

    sendFormattedResponse(res, recentInfo, format);
  } catch (error) {
    handleError(error, res, 'fetching recent updates');
  }
});

// Error handling middleware
// eslint-disable-next-line @typescript-eslint/no-unused-vars
app.use((err: Error, req: Request, res: Response, _next: NextFunction) => {
  console.error('Unhandled error:', err);
  res.status(500).json({ error: 'Internal server error' });
});

// Start server
app.listen(PORT, () => {
  console.log(`VGMdb API server listening on port ${PORT}`);
  console.log(`Try: http://localhost:${PORT}/hello`);
  console.log(`Try: http://localhost:${PORT}/album/79?format=json`);
  console.log(`Try: http://localhost:${PORT}/artist/137?format=json`);
  console.log(`Try: http://localhost:${PORT}/product/241?format=json`);
  console.log(`Try: http://localhost:${PORT}/event/138?format=json`);
  console.log(`Try: http://localhost:${PORT}/org/67?format=json`);
  console.log(`Try: http://localhost:${PORT}/albumlist/A1?format=json`);
  console.log(`Try: http://localhost:${PORT}/artistlist/A1?format=json`);
  console.log(`Try: http://localhost:${PORT}/productlist/A1?format=json`);
  console.log(`Try: http://localhost:${PORT}/orglist?format=json`);
  console.log(`Try: http://localhost:${PORT}/eventlist?format=json`);
  console.log(`Try: http://localhost:${PORT}/release/1234?format=json`);
  console.log(`Try: http://localhost:${PORT}/search?q=final+fantasy`);
  console.log(`Try: http://localhost:${PORT}/recent/albums`);
});

export default app;